Improve the logic used to update cells candidates after applying strategies
- Removing the 'next' candidates buffer within the cell object.

Instead of storing the new candidates in a 'next' candidates buffer, return
a clone of the cell with updated candidates. Then instead of updating the
cell candidates within the `grid.updateCandidates` method, simply update
it within the 'solver.resolve' method.

The idea is to deal with strategy results separately from the classic
candidate update required after the strategy results are applied.

- Do not rely on the 'grid.update' method to stop the resolve loop

It is not because the grid update does not return any result than a
strategy applied afterward cannot find more results. Stop the loop when
no strategy can return any results instead